refactor getAllDocuments, removeAll and clear
in order not to automatically load all entries in memory, the method
getAllDocuments has been transformed to getAllKeys, returning an async
Observable of all the keys in the cache.

the removeAll and clear methods have been reworked to use this in order
to remove documents in a streaming fashion, and also to be able to
notify listeners of removal in the future. Flush operation is doing more
under the covers and thus is not adapted to this feature.